workers/market_monitor/requirements.txt
# Core
redis==5.0.1
motor>=3.3.0
pymongo>=4.5.0
python-dotenv>=1.0.0

# API Clients
aiohttp==3.9.3
pydantic==2.6.3
pydantic-settings==2.2.1
PyYAML==6.0.1

# Utils
python-telegram-bot>=20.0
pytz>=2023.3
schedule>=1.2.0
websockets>=11.0.3

# Logging
python-json-logger>=2.0.0

# Testing
pytest>=6.2.5
pytest-asyncio>=0.16.0
pytest-cov>=2.12.0

# Type checking
mypy>=0.910
types-PyYAML>=5.4.0
